requirements.txt
asgiref==3.5.2
Django==3.1.5
djangorestframework==3.13.1
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
pytz==2022.1
Pillow==9.1.1
sqlparse==0.4.2
